pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Adam Goldsmith", email = "adam@adamgoldsmith.name"},
]
dependencies = [
    "requests~=2.28",
    "aprs3~=8.0"]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool]
[tool.pdm]

[tool.black]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
